# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Kobara API",
    "description": "The Kobara API for unified local and international payments, plus MonCash and NatCash withdrawals."
  },
  "servers": [
    {
      "url": "https://api.kobara.app"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Kobara API key"
      }
    },
    "schemas": {
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "pay_123456"
              },
              "reference": {
                "type": "string",
                "example": "KOB-123456"
              },
              "amount": {
                "type": "number",
                "example": 2500
              },
              "net_amount": {
                "type": "number",
                "example": 2427.5
              },
              "fee_amount": {
                "type": "number",
                "example": 72.5
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "succeeded",
                  "failed",
                  "expired",
                  "refunded"
                ],
                "example": "pending"
              },
              "environment": {
                "type": "string",
                "enum": [
                  "live"
                ]
              },
              "paid_at": {
                "type": "string",
                "nullable": true
              },
              "provider": {
                "type": "string"
              },
              "invoice_id": {
                "type": "string",
                "nullable": true
              },
              "checkout_url": {
                "type": "string",
                "format": "uri",
                "example": "https://pay.kobara.app/checkout/pay_123456"
              },
              "url": {
                "type": "string",
                "format": "uri",
                "description": "Backward-compatible alias of checkout_url",
                "example": "https://pay.kobara.app/checkout/pay_123456"
              },
              "payment_url": {
                "type": "string",
                "format": "uri",
                "description": "Backward-compatible alias of checkout_url",
                "example": "https://pay.kobara.app/checkout/pay_123456"
              },
              "paymentUrl": {
                "type": "string",
                "format": "uri",
                "description": "Backward-compatible alias of checkout_url",
                "example": "https://pay.kobara.app/checkout/pay_123456"
              }
            },
            "required": [
              "id",
              "reference",
              "amount",
              "net_amount",
              "fee_amount",
              "status",
              "environment",
              "paid_at",
              "checkout_url",
              "url",
              "payment_url",
              "paymentUrl"
            ]
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
      "PaymentCreatePayload": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Amount to charge",
            "example": 2500
          },
          "currency": {
            "type": "string",
            "enum": [
              "HTG",
              "USD"
            ],
            "default": "HTG",
            "description": "Currency of the payment",
            "example": "HTG"
          },
          "provider": {
            "type": "string",
            "enum": [
              "moncash",
              "moncash_ussd",
              "moncash_web",
              "natcash",
              "natcash_web",
              "natcash_ussd",
              "carte",
              "card",
              "paypal",
              "apple_pay",
              "google_pay",
              "crypto",
              "kobara"
            ],
            "default": "kobara",
            "description": "Payment provider (case-insensitive): moncash, natcash, card, paypal, apple_pay, google_pay, crypto, or kobara. Defaults to kobara (unified checkout page).",
            "example": "crypto"
          },
          "crypto_currency": {
            "type": "string",
            "pattern": "^[a-z0-9]{2,20}$",
            "description": "Optional NOWPayments currency ticker. Omit it to let the customer choose on the hosted crypto checkout.",
            "example": "usdttrc20"
          },
          "description": {
            "type": "string",
            "description": "Description of the payment",
            "example": "Order #12345"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "success_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to redirect after successful payment",
            "example": "https://your-site.com/success"
          },
          "cancel_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to redirect after cancelled/failed payment",
            "example": "https://your-site.com/cancel"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Key-value object for custom data",
            "example": {
              "order_id": "ORD-12345"
            }
          }
        },
        "required": [
          "amount"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the customer",
            "example": "Jean Dupont"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the customer",
            "example": "jean@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "37000000"
          }
        }
      },
...
    },
...
  },
  "paths": {
    "/v1/payments": {
      "post": {
        "summary": "Create a new payment",
        "description": "Initialize a payment transaction. Use provider \"moncash\", \"natcash\", \"card\", \"paypal\", \"apple_pay\", \"google_pay\", or \"crypto\" to preselect a method. Crypto returns a NOWPayments hosted checkout and settles the merchant balance in USD. Use \"kobara\" (default) for the unified checkout page where the customer chooses.",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Unique key for this payment attempt. Reuse it only when retrying the exact same request.",
              "example": "8f3d4e2a-93c2-4c0f-bbe0-95ab31f6d712"
            },
            "required": true,
            "description": "Unique key for this payment attempt. Reuse it only when retrying the exact same request.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment successfully initialized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (Validation error)"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}